+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/bugreporting/debuglog/upload/history/UploadHistory.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.bugreporting.uploadhistory
+package de.rki.coronawarnapp.bugreporting.debuglog.upload.history
 
 data class UploadHistory(
     val logs: List<LogUpload> = emptyList()
