+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/bugreporting/debuglog/upload/server/auth/LogUploadAuthServer.kt
@@ -1,29 +1,22 @@
-package de.rki.coronawarnapp.bugreporting.logupload.server
+package de.rki.coronawarnapp.bugreporting.debuglog.upload.server.auth
 
 import dagger.Lazy
 import dagger.Reusable
 import de.rki.coronawarnapp.appconfig.AppConfigProvider
-import de.rki.coronawarnapp.bugreporting.logupload.auth.LogUploadAuthApiV1
-import de.rki.coronawarnapp.bugreporting.logupload.auth.LogUploadOtp
 import de.rki.coronawarnapp.datadonation.safetynet.DeviceAttestation
 import de.rki.coronawarnapp.server.protocols.internal.ppdd.ElsOtp
 import de.rki.coronawarnapp.server.protocols.internal.ppdd.ElsOtpRequestAndroid
 import kotlinx.coroutines.flow.first
 import timber.log.Timber
-import java.io.File
 import java.util.UUID
 import javax.inject.Inject
 
 @Reusable
-class LogUploadServer @Inject constructor(
-    private val uploadApiProvider: Lazy<LogUploadApi>,
+class LogUploadAuthServer @Inject constructor(
     private val authApiProvider: Lazy<LogUploadAuthApiV1>,
     private val deviceAttestation: DeviceAttestation,
     private val configProvider: AppConfigProvider
 ) {
-
-    private val uploadApi: LogUploadApi
-        get() = uploadApiProvider.get()
 
     private val authApi: LogUploadAuthApiV1
         get() = authApiProvider.get()
@@ -58,10 +51,7 @@ class LogUploadServer @Inject constructor(
         )
     }
 
-    suspend fun uploadLog(otp: LogUploadOtp, log: File) {
-    }
-
     companion object {
-        private const val TAG = "LogUploadServer"
+        private const val TAG = "LogUploadOtpServer"
     }
 }
