+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/information/InformationFragmentModule.kt
@@ -6,7 +6,7 @@ import dagger.android.ContributesAndroidInjector
 import dagger.multibindings.IntoMap
 import de.rki.coronawarnapp.bugreporting.debuglog.ui.DebugLogFragmentModule
 import de.rki.coronawarnapp.bugreporting.debuglog.ui.upload.DebugLogUploadFragmentModule
-import de.rki.coronawarnapp.bugreporting.uploadhistory.ui.LogUploadHistoryModule
+import de.rki.coronawarnapp.bugreporting.debuglog.ui.upload.history.LogUploadHistoryModule
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModelFactory
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModelKey
