+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/bugreporting/BugReportSettingsTest.kt
@@ -1,8 +1,8 @@
 package de.rki.coronawarnapp.bugreporting
 
 import android.content.Context
-import de.rki.coronawarnapp.bugreporting.uploadhistory.LogUpload
-import de.rki.coronawarnapp.bugreporting.uploadhistory.UploadHistory
+import de.rki.coronawarnapp.bugreporting.debuglog.upload.history.LogUpload
+import de.rki.coronawarnapp.bugreporting.debuglog.upload.history.UploadHistory
 import de.rki.coronawarnapp.util.serialization.SerializationModule
 import io.kotest.matchers.shouldBe
 import io.mockk.MockKAnnotations
